+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v0.6.1 (2026-06-08)
+
+  * [`compressed`], [`decompress_body`]: Disable automatic decompression
+
+    Decompression is now opt-in by setting `compressed: true`.
+
 ## v0.6.0 (2026-06-08)
 
   * [`encode_body`]: Security fix for `:form_multipart` header injection